+++ src/indexes/helpers/within_group_dispercion.rs
@@ -4,7 +4,7 @@ use crate::{
     sender::{Sender, Subscriber},
 };
 
-use ndarray::{ArcArray2, Array2, ArrayView1, ArrayView2};
+use ndarray::{ Array2, ArrayView1, ArrayView2};
 use std::iter::zip;
 
 use super::raw_data::RawDataValue;
@@ -20,7 +20,7 @@ impl WGD {
     pub fn compute(
         &self,
         x: &ArrayView2<f64>,
-        y: &ArrayView2<usize>,
+        y: &ArrayView2<u32>,
         clusters_centroids: &Vec<Array2<f64>>,
     ) -> Result<Vec<Array2<f64>>, CalcError> {
         zip(y.columns(), clusters_centroids)
@@ -30,7 +30,7 @@ impl WGD {
     fn helper(
         &self,
         x: &ArrayView2<f64>,
-        y: &ArrayView1<usize>,
+        y: &ArrayView1<u32>,
         clusters_centroids: &Array2<f64>,
     ) -> Result<Array2<f64>, CalcError> {
         let (n, d) = x.dim();
